refactor: place components import with @component alias
