+++ src/components/SideMenu/SideMenuNavIcon.tsx
@@ -3,7 +3,7 @@ import { observer } from 'mobx-react';
 import { MenuAlt3 } from '@styled-icons/heroicons-outline/MenuAlt3';
 import { CloseOutline } from '@styled-icons/evaicons-outline/CloseOutline';
 
-import { MenuButton } from 'components/MenuBar';
+import { MenuButton } from '@components/MenuBar';
 import { sideMenuStore } from './state';
 
 export const SideMenuNavIcon = observer(() => {
