+++ src/pages/cv.tsx
@@ -3,7 +3,7 @@ import { graphql } from 'gatsby';
 import 'sanitize.css/sanitize.css';
 import { ArrowheadUp } from '@styled-icons/evaicons-solid/ArrowheadUp';
 
-import SEO from 'components/SEO';
+import SEO from '@components/SEO';
 import { titleWithNameAndJobTitle } from 'utils/seo';
 import { useScrollToTop } from 'hooks/useScrollToTop';
 import {
