+++ src/pages/uses.tsx
@@ -3,7 +3,7 @@ import { MDXRenderer } from 'gatsby-plugin-mdx';
 import { graphql } from 'gatsby';
 import { UsesQuery } from 'graphql-types';
 
-import Layout from 'components/Layout';
+import Layout from '@components/Layout';
 import { BlogGlobalStyle } from 'styles/blogPost';
 import { useIntl } from 'context/react-intl';
