+++ src/screens/Blog/PageTemplate.tsx
@@ -9,16 +9,16 @@ import { BlogPostProps } from '@screens/Blog/types';
 import { BlogContextProvider } from '@screens/Blog/context';
 import { useTwitterScript } from 'hooks/useTwitterScript';
 import LayoutBlog from '@screens/Blog/Layout';
-import { pageTransitionVariants, Tags, Tag } from 'components/Ui';
+import { pageTransitionVariants, Tags, Tag } from '@components/Ui';
 import { BlogGlobalStyle } from 'styles/blogPost';
 import { GlobalStyles } from 'styles/index';
-import { Container, LineDivider } from 'components/Ui';
-import { MenuBar } from 'components/MenuBar';
+import { Container, LineDivider } from '@components/Ui';
+import { MenuBar } from '@components/MenuBar';
 import { Frontmatter } from 'types';
-import SEO from 'components/SEO';
+import SEO from '@components/SEO';
 import { ThemeProvider } from 'context/theme';
 import { styled, media, SiteTheme } from '@styles/styled';
-import { SideMenu } from 'components/SideMenu';
+import { SideMenu } from '@components/SideMenu';
 import { joinSubtitleAndDescription } from 'utils/seo';
 import { AvailableTranslations } from '@screens/Blog/components/AvailableTranslations';
