+++ src/screens/Blog/components/FeaturedImage/FeaturedImage.tsx
@@ -6,7 +6,7 @@ import { styled } from '@styles/styled';
 import { useBlogContext } from '@screens/Blog/hooks/useBlogContext';
 import { Frontmatter } from 'types';
 import { useIntl } from 'context/react-intl';
-import { Container } from 'components/Ui';
+import { Container } from '@components/Ui';
 
 const StyledImg = styled(Img)`
   max-height: 600px;
