+++ src/screens/Blog/components/Header/Header.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { useBlogContext } from '@screens/Blog/hooks/useBlogContext';
 import { styled, media } from '@styles/styled';
-import { Container } from 'components/Ui';
+import { Container } from '@components/Ui';
 
 const Title = styled.h1`
   font-size: 34px;
