+++ src/screens/Blog/components/SeriesMenu/styled.ts
@@ -3,7 +3,7 @@ import { motion } from 'framer-motion';
 
 import { styled, media } from '@styles/styled';
 import { pxToRem } from 'styles/blogPost';
-import { Card as DefaultCard } from 'components/Ui';
+import { Card as DefaultCard } from '@components/Ui';
 
 export const Card = styled(DefaultCard)`
   padding: 0;
