+++ src/types/styled-components.d.ts
@@ -1,7 +1,7 @@
 // import original module declarations
 import 'styled-components';
 import { SiteTheme } from '@styles/styled';
-import { CvTheme } from 'components/CV/styles';
+import { CvTheme } from '@components/CV/styles';
 
 // and extend them!
 declare module 'styled-components' {
